Jon dev

+++ src/edu/worcester/cs499summer2012/comparator/TaskDateDueComparator.java
@@ -60,7 +60,7 @@ public class TaskDateDueComparator implements Comparator<Task> {
 				// Ex. LHS -> date 5000ms (earlier)
 				//     RHS -> date 6000ms (later)
 				//     LHS - RHS = 5000 - 6000 = -1000 = LHS ordered first
-				return lhs.getDateDue() - rhs.getDateDue();
+				return (int) (lhs.getDateDue() - rhs.getDateDue());
 			}
 		}
 	}
