+++ src/edu/worcester/cs499summer2012/comparator/TaskDateModifiedComparator.java
@@ -42,6 +42,6 @@ public class TaskDateModifiedComparator implements Comparator<Task> {
 		// Ex. LHS -> date 5000ms (earlier)
 		//     RHS -> date 6000ms (later)
 		//     LHS - RHS = 5000 - 6000 = -1000 = LHS ordered first
-		return lhs.getDateModified() - rhs.getDateModified();
+		return (int) (lhs.getDateModified() - rhs.getDateModified());
 	}
 }
